refactor(ui): remove internal system_ui legacy bridge
