+++ ui/crates/system_ui/src/lib.rs
@@ -9,9 +9,8 @@
 #![warn(missing_docs, rustdoc::broken_intra_doc_links)]
 #![allow(missing_docs)]
 
+mod foundation;
 mod icon;
-#[path = "primitives/mod.rs"]
-mod legacy_primitives;
 mod origin_components;
 mod origin_primitives;
 mod origin_tokens;
@@ -33,19 +32,19 @@ pub mod tokens {
 /// Prefer importing from this module in app crates so primitive usage stays consistent and review
 /// diffs do not churn on long individual import lists.
 pub mod prelude {
+    pub use crate::components::StepStatus;
     pub use crate::components::{
         AppShell, Button, DisclosurePanel, IconButton, StatusBar, StatusBarItem, StepFlow,
         StepFlowActions, StepFlowHeader, StepFlowStep, ToggleRow, WindowControls, WindowFrame,
         WindowTitleBar,
     };
-    pub use crate::legacy_primitives::StepStatus;
     pub use crate::primitives::{
         ButtonShape, ButtonSize, ButtonVariant, Center, CheckboxField, Cluster, Elevation,
         FieldVariant, Grid, Heading, Icon, IconName, IconSize, Inline, Inset, Layer, LayoutAlign,
         LayoutGap, LayoutJustify, LayoutPadding, ListSurface, Panel, ResizeHandleRegion, Stack,
-        Surface, SurfaceVariant, TerminalLine, TerminalPrompt, TerminalSurface,
-        TerminalTranscript, Text, TextField, TextRole, TextTone, TitlebarRegion, Viewport,
-        WindowBody, WindowControlButton, WindowSurface, WindowTitle,
+        Surface, SurfaceVariant, TerminalLine, TerminalPrompt, TerminalSurface, TerminalTranscript,
+        Text, TextField, TextRole, TextTone, TitlebarRegion, Viewport, WindowBody,
+        WindowControlButton, WindowSurface, WindowTitle,
     };
     pub use crate::tokens::baseline_style_id;
 }
