+++ ui/crates/system_ui/src/origin_components/mod.rs
@@ -1,16 +1,13 @@
 mod actions;
 mod navigation;
-mod overlays;
 mod shell;
 mod windowing;
 
-pub use crate::legacy_primitives::StepStatus;
 pub use actions::{Button, IconButton};
 pub use navigation::{
     DisclosurePanel, MenuItem, MenuSeparator, MenuSurface, StatusBar, StatusBarItem, StepFlow,
-    StepFlowActions, StepFlowHeader, StepFlowStep, Toolbar, ToggleRow,
+    StepFlowActions, StepFlowHeader, StepFlowStep, StepStatus, ToggleRow, Toolbar,
 };
-pub use overlays::{ContextMenu, Launcher, ModalOverlay};
 pub use shell::{
     AppShell, ClockButton, DesktopBackdrop, DesktopIcon, DesktopIconGrid, DesktopWindowLayer,
     SystemTray, Taskbar, TaskbarButton, TaskbarOverflowButton, TaskbarSection, TrayButton,
