+++ ui/crates/system_ui/src/origin_primitives/mod.rs
@@ -1,25 +1,25 @@
-pub use crate::icon::{Icon, IconName, IconSize};
-pub use crate::legacy_primitives::{
-    ButtonShape, ButtonSize, ButtonVariant, Cluster, Elevation, FieldVariant, Grid, LayoutAlign,
-    LayoutGap, LayoutJustify, LayoutPadding, ListSurface, Panel, Stack, SurfaceVariant, TextRole,
-    TextTone,
+pub use crate::foundation::{
+    ButtonShape, ButtonSize, ButtonVariant, Elevation, FieldVariant, LayoutAlign, LayoutGap,
+    LayoutJustify, LayoutPadding, SurfaceVariant, TextRole, TextTone,
 };
-pub use crate::legacy_primitives::{
-    CompletionItem, CompletionList, DataTable, Surface, TerminalLine, TerminalPrompt,
-    TerminalSurface, TerminalTranscript,
+pub use crate::icon::{Icon, IconName, IconSize};
+pub use data::{
+    CompletionItem, CompletionList, DataTable, TerminalLine, TerminalPrompt, TerminalSurface,
+    TerminalTranscript,
 };
 
 mod controls;
+mod data;
 mod layout;
 mod surfaces;
 mod typography;
 mod window;
 
 pub use controls::{CheckboxField, TextField};
+pub use data::{Cluster, Grid, ListSurface, Panel, Stack, Surface};
 pub use layout::{Center, Inline, Inset};
 pub use surfaces::{Layer, Viewport};
 pub use typography::{Heading, Text};
 pub use window::{
-    ResizeHandleRegion, TitlebarRegion, WindowBody, WindowControlButton, WindowSurface,
-    WindowTitle,
+    ResizeHandleRegion, TitlebarRegion, WindowBody, WindowControlButton, WindowSurface, WindowTitle,
 };
