+++ ui/crates/system_ui/src/origin_primitives/surfaces.rs
@@ -1,6 +1,6 @@
 use leptos::*;
 
-use crate::legacy_primitives::{merge_layout_class, Elevation, LayoutPadding};
+use crate::foundation::{merge_layout_class, Elevation, LayoutPadding};
 
 #[component]
 pub fn Layer(
