+++ ui/crates/system_ui/tests/public_api.rs
@@ -3,4 +3,5 @@ fn reduced_prelude_public_api() {
     let cases = trybuild::TestCases::new();
     cases.pass("tests/ui/prelude_allowed.rs");
     cases.compile_fail("tests/ui/prelude_rejects_badge.rs");
+    cases.compile_fail("tests/ui/legacy_module_hidden.rs");
 }
