+++ ui/crates/system_ui/tests/ui/legacy_module_hidden.rs
@@ -1,0 +1,5 @@
+use system_ui::legacy_primitives::StepStatus;
+
+fn main() {
+    let _ = StepStatus::Current;
+}
